Se cambia el tipo de dato de int a long para el perfil.

+++ CotsWS/src/main/java/mx/morena/negocio/servicios/ICotService.java
@@ -6,6 +6,6 @@ import mx.morena.negocio.dto.CotDTO;
 import mx.morena.negocio.exception.CotException;
 
 public interface ICotService {
-	public String save(CotDTO cot, int perfil) throws CotException;
-	public String asignarSecciones(List<Long> idSecciones, Long idCot, int perfil) throws CotException;
+	public String save(CotDTO cot, long perfil) throws CotException;
+	public String asignarSecciones(List<Long> idSecciones, Long idCot, long perfil) throws CotException;
 }
